misc additional cleanup and lots of docs
